requirements.txt
django>=4.2.0
djangorestframework==3.15.2
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
pytest==8.3.4
psycopg2-binary==2.9.10
pillow==11.0.0
python-dotenv==1.0.1

django-celery-beat==2.7.0
django-celery-results==2.5.1

gunicorn==22.0.0
whitenoise==6.9.0
django-cors-headers==4.7.0
pytest-django==4.10.0
djangorestframework-simplejwt==5.5.0
selenium==4.15.2
webdriver-manager==4.0.1
pytest-selenium==4.0.2

